add settings parser and restructure files. also change dat file read in 'insert' to operate byte by byte
